rework file structure for requests and add logout request
