+++ app/Http/Controllers/Api/Auth/LoginDomain/LoginController.php
@@ -3,7 +3,7 @@
 namespace App\Http\Controllers\Api\Auth\LoginDomain;
 
 use Illuminate\Http\Request;
-use App\Http\Requests\LoginRequest;
+use App\Http\Requests\Auth\LoginLogoutRequest;
 use App\Http\Controllers\Controller;
 use App\Http\Controllers\Api\Auth\LoginDomain\Contracts\LoginContract;
 
@@ -16,7 +16,7 @@ class LoginController extends Controller
         $this->loginRetriever = $loginContract;
     }
 
-    public function login(LoginRequest $request)
+    public function login(LoginLogoutRequest $request)
     {
         return $this->loginRetriever->login($request);
     }
