+++ app/Http/Controllers/Api/Auth/LogoutDomain/LogoutController.php
@@ -2,6 +2,7 @@
 namespace App\Http\Controllers\Api\Auth\LogoutDomain;
 
 use Illuminate\Http\Request;
+use App\Http\Requests\Auth\LoginLogoutRequest;
 use App\Http\Controllers\Controller;
 use App\Http\Controllers\Api\Auth\LogoutDomain\Contracts\LogoutContract;
 
@@ -14,7 +15,7 @@ class LogoutController extends Controller
         $this->logoutRetriever = $logoutContract;
     }
 
-    public function logout(Request $request)
+    public function logout(LoginLogoutRequest $request)
     {
         return $this->logoutRetriever->logout($request);
     }
