+++ app/Http/Controllers/Api/Auth/RegisterDomain/RegisterController.php
@@ -2,7 +2,7 @@
 namespace App\Http\Controllers\Api\Auth\RegisterDomain;
 
 use Illuminate\Http\Request;
-use App\Http\Requests\RegisterRequest;
+use App\Http\Requests\Auth\RegisterRequest;
 use App\Http\Controllers\Controller;
 use App\Http\Controllers\Api\Auth\RegisterDomain\Contracts\RegisterContract;
