+++ app/Http/Requests/Auth/LoginLogoutRequest.php
@@ -1,10 +1,10 @@
 <?php
 
-namespace App\Http\Requests;
+namespace App\Http\Requests\Auth;
 
 use Illuminate\Foundation\Http\FormRequest;
 
-class RegisterRequest extends FormRequest
+class LoginLogoutRequest extends FormRequest
 {
     /**
      * Determine if the user is authorized to make this request.
@@ -24,19 +24,17 @@ class RegisterRequest extends FormRequest
     public function rules()
     {
         return [
-            'name' => 'required',
-            'email' => 'required|email|unique:users,email',
-            'password' => 'required|min:6|confirmed'
+            'username' => 'required|email|unique:users,email',
+            'password' => 'required|confirmed'
         ];
     }
 
     public function messages()
     {
         return [
-            'name.required' => 'Name is required!',
-            'body.required' => 'Email is required!',
-            'password.required' => 'Password is required!',
-            'password.min' => 'Password must be 6 characters long!'
+            'username.required' => 'Username is required!',
+            'username.email' => 'Username must be an email!',
+            'password.required' => 'Password is required!'
         ];
     }
 }
